Updated comment to describe behaviour, updated custom validator class name and error checks
